shotgun b64 encoding now uses size_t for lengths to avoid overflows
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@75858 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ PROTO/shotgun/src/lib/cencode.c
@@ -23,7 +23,7 @@ char base64_encode_value(char value_in)
         return encoding[(int)value_in];
 }
 
-int base64_encode_block(const char* plaintext_in, int length_in, char* code_out, base64_encodestate* state_in)
+int base64_encode_block(const char* plaintext_in, size_t length_in, char* code_out, base64_encodestate* state_in)
 {
         const char* plainchar = plaintext_in;
         const char* const plaintextend = plaintext_in + length_in;
